Improve ``run_shell`` command functionality and implementation.
It can be used with no arguments, or as follows::

  run_shell good "my message here"
  run_shell bad "my message here"

When used the above way, it will print a ``good_msg`` or
``bad_msg`` you specify and then start the rescue shell.
This helps to streamline the ``linuxrc`` code.

Also add a ``good_msg`` saying "Starting rescue shell:"
which makes it clear that the shell started intentionally
and not as a bug.