New argument parser for ramdisk (work in progress).
I have implemented a new argument parser so I am not forced to use the
hideous "argparse".

There are some things that are not yet working. You can't specify the
kernel to build. It will always use the ``/usr/src/linux`` symlink. This
will be fixed soon. I plan to improve things here as it was previously
too hard to select a different kernel. So -- this part was scheduled for
demolition anyway. Will get to it soon.